+++ src/twebsocket.cpp
@@ -48,7 +48,7 @@ void TWebSocket::close()
 
 void TWebSocket::sendText(const QString &text)
 {
-    tSystemDebug("sendText  text len:%d  (pid:%d)", text.length(), QCoreApplication::applicationPid());
+    tSystemDebug("sendText  text len:%d  (pid:%d)", text.length(), (int)QCoreApplication::applicationPid());
     TAbstractWebSocket::sendText(text);
 }
